[REFACTOR] Restructure folder and clean up unused components
