+++ app/components/about/SectionContainer.tsx
