+++ app/components/help/HelpSection.tsx
