+++ app/components/home/AdvantagesSection.tsx
