+++ app/components/home/HeroSection.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import ButtonBase from "../components/ButtonBase";
+import ButtonBase from "../ButtonBase";
 
 export default function HeroSection() {
   return (
