+++ app/components/home/MapGallery.tsx
