+++ app/page.tsx
@@ -1,11 +1,11 @@
 import "../styles/globals.css";
 import Footer from "./components/Footer";
-import HeroSection from "./components/HeroSection";
-import WhyPantauTularSection from "./components/WhyPantauTularSection";
-import FeaturesSection from "./components/AdvantagesSection";
-import AboutSection from "./components/AboutSection";
-import HelpSection from "./components/HelpSection";
-import MapGallery from "./components/MapGallery";
+import HeroSection from "./components/home/HeroSection";
+import WhyPantauTularSection from "./components/home/WhyPantauTularSection";
+import FeaturesSection from "./components/home/AdvantagesSection";
+import AboutSection from "./components/home/AboutSection";
+import HelpSection from "./components/home/HelpSection";
+import MapGallery from "./components/home/MapGallery";
 import Navbar from "./components/Navbar";
 
 export default function Home() {
